rename IAggregateRoot into IHasKey
