+++ Untech.Practices.sln
@@ -6,6 +6,7 @@ MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{839A588F-BFF8-4B49-BD50-235245F9B16A}"
 	ProjectSection(SolutionItems) = preProject
 		.editorconfig = .editorconfig
+		src\Directory.build.props = src\Directory.build.props
 	EndProjectSection
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Untech.Practices.Collections", "src\Untech.Practices.Collections\Untech.Practices.Collections.csproj", "{9E8D6A84-28EC-4DDB-B70F-93A9000C636E}"
