+++ examples/Shared.Infrastructure/Data/UserScopedGenericDataStorage.cs
@@ -10,7 +10,7 @@ using Untech.Practices.UserContext;
 namespace Shared.Infrastructure.Data
 {
 	public class UserScopedGenericDataStorage<T, TTable, TMapper> : IDataStorage<T>
-		where T : IAggregateRoot
+		where T : IHasKey
 		where TTable : class, IUserScopedDao
 		where TMapper : struct, IDaoMapper<T, TTable>
 	{
@@ -29,7 +29,7 @@ namespace Shared.Infrastructure.Data
 			{
 				var dao = await GetMyItems(dataContext)
 						.FirstOrDefaultAsync(n => n.Key == key, cancellationToken)
-					?? throw new AggregateRootNotFoundException(key);
+					?? throw new ItemNotFoundException(key);
 
 				return FromDao(dao);
 			}
