+++ src/Untech.Practices.DataStorage.Abstractions/IDataStorage.cs
@@ -4,20 +4,20 @@ using System.Threading.Tasks;
 namespace Untech.Practices.DataStorage
 {
 	/// <summary>
-	///     Defines async CRUD methods for entities that inherit <see cref="IAggregateRoot{TKey}" />.
+	///     Defines async CRUD methods for entities that inherit <see cref="IHasKey{TKey}" />.
 	/// </summary>
 	/// <typeparam name="T">Type of the entity.</typeparam>
 	/// <typeparam name="TKey">Type of the entity key.</typeparam>
 	public interface IDataStorage<T, TKey>
-		where T : IAggregateRoot<TKey>
+		where T : IHasKey<TKey>
 	{
 		/// <summary>
 		///     Finds entity asynchronously with a specified <paramref name="key" /> in the store.
 		/// </summary>
 		/// <param name="key">Key of the entity to find.</param>
 		/// <param name="cancellationToken">Task cancellation token.</param>
-		/// <returns>Entity that was found; otherwise throws <see cref="AggregateRootNotFoundException" />.</returns>
-		/// <exception cref="AggregateRootNotFoundException">Entity with <paramref name="key" /> cannot be found.</exception>
+		/// <returns>Entity that was found; otherwise throws <see cref="ItemNotFoundException" />.</returns>
+		/// <exception cref="ItemNotFoundException">Entity with <paramref name="key" /> cannot be found.</exception>
 		Task<T> GetAsync(TKey key, CancellationToken cancellationToken = default);
 
 		/// <summary>
@@ -34,7 +34,7 @@ namespace Untech.Practices.DataStorage
 		/// <param name="entity">Entity to create.</param>
 		/// <param name="cancellationToken">Task cancellation token.</param>
 		/// <returns>Entity that was created in the store.</returns>
-		/// <exception cref="AggregateRootNotFoundException">Entity with <paramref name="key" /> cannot be found.</exception>
+		/// <exception cref="ItemNotFoundException">Entity with <paramref name="key" /> cannot be found.</exception>
 		Task<T> UpdateAsync(T entity, CancellationToken cancellationToken = default);
 
 		/// <summary>
@@ -43,16 +43,16 @@ namespace Untech.Practices.DataStorage
 		/// <param name="entity">Entity to delete.</param>
 		/// <param name="cancellationToken">Task cancellation token.</param>
 		/// <returns>True when deleted successfully; otherwise false.</returns>
-		/// <exception cref="AggregateRootNotFoundException">Entity with <paramref name="key" /> cannot be found.</exception>
+		/// <exception cref="ItemNotFoundException">Entity with <paramref name="key" /> cannot be found.</exception>
 		Task<bool> DeleteAsync(T entity, CancellationToken cancellationToken = default);
 	}
 
 	/// <summary>
-	///     Defines async CRUD methods for entities with <see cref="int" /> key that inherits <see cref="IAggregateRoot" />.
+	///     Defines async CRUD methods for entities with <see cref="int" /> key that inherits <see cref="IHasKey" />.
 	/// </summary>
 	/// <typeparam name="T"></typeparam>
 	public interface IDataStorage<T> : IDataStorage<T, int>
-		where T : IAggregateRoot
+		where T : IHasKey
 	{
 	}
 }
