+++ src/Untech.Practices.DataStorage.Linq2Db/IEntityMapper.cs
@@ -1,7 +1,8 @@
 namespace Untech.Practices.DataStorage.Linq2Db
 {
 	public interface IEntityMapper<T, TDao, TKey>
-		where T : IAggregateRoot<TKey>
+		where T : IHasKey<TKey>
+		where TDao: IHasKey<TKey>
 	{
 		T Map(TDao value);
